requirements.txt
certifi==2019.11.28
paho-mqtt==1.5.0
